Updates related to the leading dimensions.

+++ src/aalloc.c
@@ -24,7 +24,7 @@ int salloc2_(const fnat m[static restrict 1], const fnat n[static restrict 1], f
 #endif /* ?_OPENMP */
     for (fnat j = 0u; j < *n; ++j) {
       register const VS z = _mm512_setzero_ps();
-      float *const Aj = *A + j * (*ldA);
+      float *const Aj = *A + j * (size_t)(*ldA);
       for (fnat i = 0u; i < *ldA; i += VSL)
         _mm512_store_ps((Aj + i), z);
 #ifdef _OPENMP
@@ -60,7 +60,7 @@ int dalloc2_(const fnat m[static restrict 1], const fnat n[static restrict 1], d
 #endif /* ?_OPENMP */
     for (fnat j = 0u; j < *n; ++j) {
       register const VD z = _mm512_setzero_pd();
-      double *const Aj = *A + j * (*ldA);
+      double *const Aj = *A + j * (size_t)(*ldA);
       for (fnat i = 0u; i < *ldA; i += VDL)
         _mm512_store_pd((Aj + i), z);
 #ifdef _OPENMP
@@ -102,7 +102,7 @@ int calloc2_(const fnat m[static restrict 1], const fnat n[static restrict 1], f
 #endif /* ?_OPENMP */
     for (fnat j = 0u; j < *n; ++j) {
       register const VS z = _mm512_setzero_ps();
-      float complex *const Aj = *A + j * (*ldA);
+      float complex *const Aj = *A + j * (size_t)(*ldA);
       for (fnat i = 0u; i < *ldA; i += VSL_2)
         _mm512_store_ps((Aj + i), z);
 #ifdef _OPENMP
@@ -154,7 +154,7 @@ int zalloc2_(const fnat m[static restrict 1], const fnat n[static restrict 1], d
 #endif /* ?_OPENMP */
     for (fnat j = 0u; j < *n; ++j) {
       register const VD z = _mm512_setzero_pd();
-      double complex *const Aj = *A + j * (*ldA);
+      double complex *const Aj = *A + j * (size_t)(*ldA);
       for (fnat i = 0u; i < *ldA; i += VDL_2)
         _mm512_store_pd((Aj + i), z);
 #ifdef _OPENMP
