+++ src/dznrmx.h
@@ -9,7 +9,7 @@
   for (fnat j = 0u; j < *n; ++j) {                                                     \
     register const VD inf = _mm512_set1_pd(HUGE_VAL);                                  \
     register VD x = _mm512_set1_pd(-HUGE_VAL);                                         \
-    const double *const Aj = (A) + j * (*(ldA));                                       \
+    const double *const Aj = (A) + j * (size_t)(*(ldA));                               \
     for (fnat i = 0u; i < *m; i += VDL)                                                \
       x = _mm512_max_pd(_mm512_min_pd(_mm512_abs_pd(_mm512_load_pd(Aj + i)), inf), x); \
     y = fmax(_mm512_reduce_max_pd(x), y);                                              \
@@ -17,7 +17,7 @@
 #else /* !_OPENMP */
 #define DZNRMX_LOOP(A,ldA)                                                                     \
   for (fnat j = 0u; j < *n; ++j) {                                                             \
-    const double *const Aj = (A) + j * (*(ldA));                                               \
+    const double *const Aj = (A) + j * (size_t)(*(ldA));                                       \
     for (fnat i = 0u; i < *m; i += VDL) {                                                      \
       x = _mm512_max_pd(_mm512_min_pd(_mm512_abs_pd(_mm512_load_pd(Aj + i)), inf), x); VDP(x); \
     }                                                                                          \
