+++ src/zsplit.c
@@ -33,9 +33,9 @@ int zsplit_(const fnat m[static restrict 1], const fnat n[static restrict 1], co
 #pragma omp parallel for default(none) shared(m,n,A,ldA,Ar,ldAr,Ai,ldAi) reduction(max:t)
   for (fnat j = 0u; j < *n; ++j) {
     register const VI idx = _mm512_set_epi64(7, 5, 3, 1, 6, 4, 2, 0);
-    const double complex *const Aj = A + j * (*ldA);
-    double *const Arj = Ar + j * (*ldAr);
-    double *const Aij = Ai + j * (*ldAi);
+    const double complex *const Aj = A + j * (size_t)(*ldA);
+    double *const Arj = Ar + j * (size_t)(*ldAr);
+    double *const Aij = Ai + j * (size_t)(*ldAi);
     for (fnat i = 0u; i < *m; i += VDL_2) {
       register const VD ri = _mm512_permutexvar_pd(idx, _mm512_load_pd(Aj + i));
       _mm256_store_pd((Arj + i), _mm512_extractf64x4_pd(ri, 0x00u));
@@ -49,9 +49,9 @@ int zsplit_(const fnat m[static restrict 1], const fnat n[static restrict 1], co
   register const VI idx = _mm512_set_epi64(7, 5, 3, 1, 6, 4, 2, 0);
 
   for (fnat j = 0u; j < *n; ++j) {
-    const double complex *const Aj = A + j * (*ldA);
-    double *const Arj = Ar + j * (*ldAr);
-    double *const Aij = Ai + j * (*ldAi);
+    const double complex *const Aj = A + j * (size_t)(*ldA);
+    double *const Arj = Ar + j * (size_t)(*ldAr);
+    double *const Aij = Ai + j * (size_t)(*ldAi);
     for (fnat i = 0u; i < *m; i += VDL_2) {
       register const VD ri = _mm512_permutexvar_pd(idx, _mm512_load_pd(Aj + i)); VDP(ri);
       _mm256_store_pd((Arj + i), _mm512_extractf64x4_pd(ri, 0x00u));
